powerproxy-spring-boot-starter/src/main/kotlin/io/powerproxy/integrations/boot/PowerProxyUserArgumentResolver.kt
package io.powerproxy.integrations.boot

import org.springframework.core.MethodParameter
import org.springframework.security.core.context.SecurityContextHolder
import org.springframework.security.oauth2.server.resource.authentication.JwtAuthenticationToken
import org.springframework.web.bind.support.WebDataBinderFactory
import org.springframework.web.context.request.NativeWebRequest
import org.springframework.web.method.support.HandlerMethodArgumentResolver
import org.springframework.web.method.support.ModelAndViewContainer

class PowerProxyUserArgumentResolver : HandlerMethodArgumentResolver {
    override fun supportsParameter(parameter: MethodParameter): Boolean =
        parameter.parameterType == User::class.java

    override fun resolveArgument(
        parameter: MethodParameter,
        mavContainer: ModelAndViewContainer?,
        webRequest: NativeWebRequest,
        binderFactory: WebDataBinderFactory?
    ): Any? = when (val authentication = SecurityContextHolder.getContext().authentication) {
        is JwtAuthenticationToken -> JwtUser(authentication)
        else -> null
    }
}
